% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedcontrasts.R
\name{contr.poly.weighted}
\alias{contr.poly.weighted}
\title{Orthogonal polynomial contrast matrices for unbalanced data}
\usage{
contr.poly.weighted(f, weights = NULL, width = 1)
}
\arguments{
\item{f}{a factor}

\item{weights}{optional vector of weights. If not supplied, weights
are calculated from \code{f}}

\item{width}{width of group variable (only affects the linear term)}
}
\value{
contrast matrix
}
\description{
Similar to \link[stats:contrast]{contr.poly}, but will provide orthogonal contrasts
for unbalanced data. By specifying the `width` argument correctly
the function will also return a contrast matrix with an interpretable
linear term.
}
