% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prostate}
\alias{prostate}
\title{Prostate cancer deaths for nonwhites in the U.S., 1935-1970}
\format{
A data frame with 49 rows and 5 variables:
\describe{
    \item{\code{age}}{Age group}
    \item{\code{period}}{Period group}
    \item{\code{cohort}}{Cohort group}
    \item{\code{deaths}}{Number of deaths}
    \item{\code{population}}{Population size}
}
}
\source{
Holford, Theodore R. (1983).
 The Estimation of Age, Period and Cohort Effects for Vital Rates.
 Biometrics 39(2): 311-324.
}
\usage{
prostate
}
\description{
Dataset used by Holford (1983), Table 2
}
\keyword{datasets}
