% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wordsum}
\alias{wordsum}
\title{Data from apcR package.}
\format{
An object of class \code{data.frame} with 23824 rows and 9 columns.
}
\source{
Fosse, Ethan and Christopher Winship. 2019.
  Analyzing Age-Period-Cohort Data: A Review and Critique,
  Annual Review of Sociology 45:467–92.
}
\usage{
wordsum
}
\description{
Dataset used by Fosse and Winship 2019
}
\keyword{datasets}
